Fix signed/unsigned mismatch
(cherry picked from commit 4f9dbeb22261424248f64ec60d41e50c231d7817)

+++ hpx/parallel/algorithms/find.hpp
@@ -73,7 +73,7 @@ namespace hpx { namespace parallel { HPX_INLINE_NAMESPACE(v1)
                         [=](std::vector<hpx::future<void> > &&) mutable -> InIter
                         {
                             std::size_t find_res = tok.get_data();
-                            if(find_res != count)
+                            if(find_res != static_cast<std::size_t>(count))
                                 std::advance(first, find_res);
                             else
                                 first = last;
@@ -202,7 +202,7 @@ namespace hpx { namespace parallel { HPX_INLINE_NAMESPACE(v1)
                         [=](std::vector<hpx::future<void> > &&) mutable -> FwdIter
                         {
                             std::size_t find_res = tok.get_data();
-                            if(find_res != count)
+                            if(find_res != static_cast<std::size_t>(count))
                                 std::advance(first, find_res);
                             else
                                 first = last;
@@ -351,7 +351,7 @@ namespace hpx { namespace parallel { HPX_INLINE_NAMESPACE(v1)
                         [=](std::vector<hpx::future<void> > &&) mutable -> FwdIter
                         {
                             std::size_t find_res = tok.get_data();
-                            if(find_res != count)
+                            if(find_res != static_cast<std::size_t>(count))
                                 std::advance(first, find_res);
                             else
                                 first = last;
